cli: extract status pkg
